+++ verik-compiler/src/main/kotlin/io/verik/compiler/main/VerikReport.kt
@@ -18,5 +18,6 @@ package io.verik.compiler.main
 
 class VerikReport {
 
+    var entryPoints: List<String> = listOf()
     var counts: List<Pair<String, Int>> = listOf()
 }